Update missionplanner.cpp
The locations file was previously a hard-coded path. It was changed to be dependent on the ROS package path during an end of the year code review. The package path is set after the locations file has already been loaded. Therefore, the loadFavorites method needs to be called only after the package path is known. The method was renamed to better represent this new responsibility.